+++ android/app/src/main/kotlin/ir/smartday/smart_day_planner/MainActivity.kt
@@ -1,0 +1,5 @@
+package ir.smartday.smart_day_planner
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()